refactor(rename): rename file and component name to order table
